requirements.txt
flask==3.0.0
flask-sqlalchemy
psycopg2-binary
python-dotenv
google-generativeai
